+++ pam/packages/ui/src/index.ts
@@ -53,6 +53,7 @@ export {
   StarIcon,
   HomeIcon,
   MeIcon,
+  MeIconFilled,
   PeopleIcon,
   PhoneIcon,
   PlacesIcon,